MCR-3640 refactored test code
